ci(gw): improve python bindings build process (#12)
## Motivation

The current way we build python packages is not reproducible out of
github actions.

## Solution

This PR allows to build python wheel also locally, and deterministically install
the required python versions with uv and then use them with maturin. We are
not bounded to the python installed in the system and we don't neet to use
manylinux containers.

Signed-off-by: Mauro Sardara <msardara@cisco.com>