feat(tables): do not require Default/Clone traits for elements stored in pool (#97)
## Motivation

At the moment, the pool implementation in the gateway requires the underlying contained type to implement
the `Default` and `Clone` traits. While this is acceptable in general, it does not work well with types like Box,
that cannot be cloned and mey not implement the Default trait.

## Solution

Enable any type to be stored in the pool data structure.