feat(data-plane/service): first draft of session layer (#106)
BREAKING CHANGE: This change breaks the python binding interface.

## Motivation

We need a reliability layer that is able to handle message recovery in case the message sent across
multiple gateways is lost.

## Solution

This PRs provides a first implementation of the session layer and the fire and forget delivery, that does not ensure any reliability.

Signed-off-by: Mauro Sardara <msardara@cisco.com>