fix: remove request-reply session type (#416)
BREAKING CHANGE: remove of one session type

# Description

We remove the request-reply session, as request-reply
semantics can be easily implemented in the application layer
if the application requires a reply.

Most of the time, applications require sessions that can handle
bidirectional streaming (e.g. fire-and-forget for anycast and pubsub
for multicast), on top of which we can easily build-up a request-reply
semantic. No need to implement a specific session type only for this.

We will still keep the high level API request-reply in the python
bindings, but the underlying session will not be a request-reply
only session, but rather a bidirectional streaming session.

Signed-off-by: Mauro Sardara <msardara@cisco.com>