feat(srpc): Improve SRPCChannel usability
This makes the SRPCChannel kick off the preparation in the background
and then lazyily awaits that preparation task to be completed.

This allows the SRPCChannel creation to remain a sync interface which
makes it more compatible with things like the A2A SDK but still prepare
itself using async tasks from the SLIM python bindings.