feat(slimrpc): Separate Channel and ChannelFactory
This allows you to create a ChannelFactory separately from a Channel,
the SLIM binding instance will belong to the ChannelFactory and only the
session/connection to a particular remote will be owned by the Channel.

This prevents applications which use a lot of Channels / dynamically
create channels from having to join the SLIM network many times with
different unique identities, instead the Channels reuse the slim
connection from the ChannelFactory.