refactor(service): session files in separate module (#695)
# Description

The service crate contains the session layer
and the public interface for applications using the SLIM APIs,
but it should ideally only contains the public interface for
SLIM applications. Session layer should live in another crate.

To ease up the refactor in the future, we need to separate
the sesion layer from the rest by creating a specific session module.

Signed-off-by: Mauro Sardara <msardara@cisco.com>