feat(python/bindings): create a unique SLIM data-plane instance per process by default (#819)
# Description

Currently we create a dataplane instance for each Slim instance
we build in the python bindings. This is very practical for testing, as
we can build arbitrary topologies directly from the python bindings,
but it is challenging when building real application:
 - we cannot create 2 different apps on top of the same data-plane,
    as each app would have its own dataplane
 - If we destroy and re-create a Slim instance, we need to re-establish
    all the connection with the remote Slim

To overcome this, from now the data-plane is by default instantiated
statically at program initialization.

We can still chose to instantiate it together with the Slim python
class, but this is not recommended for real apps.

Fixes: #820

## Type of Change

- [ ] Bugfix
- [x] New Feature
- [ ] Breaking Change
- [ ] Refactor
- [ ] Documentation
- [ ] Other (please describe)

## Checklist

- [x] I have read the [contributing
guidelines](/agntcy/repo-template/blob/main/CONTRIBUTING.md)
- [x] Existing issues have been referenced (where applicable)
- [x] I have verified this change is not present in other open pull
requests
- [x] Functionality is documented
- [x] All code style checks pass
- [x] New code contribution is covered by automated tests
- [x] All new and existing tests pass

---------

Signed-off-by: Mauro Sardara <msardara@cisco.com>