feat(python-bindings): Remove Py prefix from the python names
In rust we prefix types exposed to Python with "Py" to disiquish them
from the pure rust types. By default pyo3 uses the rust names as the
python names but this exposes the "Py" prefix to python users which is a
bit weird. Like "slim_bindings.PyName" instead of "slim_bindings.Name".

This commit adds a name override the pyclass / pyo3 annotations to
expose python friendly names in the bindings.

Signed-off-by: Sam Betts <1769706+Tehsmash@users.noreply.github.com>