fix(session): correctly remove routes on session close (#1039)
# Description

This PR fixes session route cleanup so that routes are correctly removed
when sessions close. In particular the fix introduces a ref count for
each route so that if multiple sessions use the same route, the route is
removed only if the ref count goes to 0.

The PR improves the testing for P2P application to test 2 sessions in
parallel.

fix https://github.com/agntcy/slim/issues/1033

## Type of Change

- [x] Bugfix
- [ ] New Feature
- [ ] Breaking Change
- [ ] Refactor
- [ ] Documentation
- [ ] Other (please describe)

## Checklist

- [x] I have read the [contributing
guidelines](/agntcy/repo-template/blob/main/CONTRIBUTING.md)
- [x] Existing issues have been referenced (where applicable)
- [x] I have verified this change is not present in other open pull
requests
- [x] Functionality is documented
- [x] All code style checks pass
- [x] New code contribution is covered by automated tests
- [x] All new and existing tests pass

---------

Signed-off-by: Michele Papalini <micpapal@cisco.com>
Signed-off-by: Michele Papalini <49271675+micpapal@users.noreply.github.com>
Co-authored-by: Mauro Sardara <msardara@cisco.com>