fix: p2p integration test (#1049)
# Description

This PR fixes an issue with parallel p2p session tests. Now, if the test
is configured to connect to multiple remote applications, it runs
multiple clients for each name to test the discovery process. Otherwise,
only one client is spawned so that the name of the remote endpoint is
known upfront.

Fix #1048 

## Type of Change

- [x] Bugfix
- [ ] New Feature
- [ ] Breaking Change
- [ ] Refactor
- [ ] Documentation
- [ ] Other (please describe)

## Checklist

- [x] I have read the [contributing
guidelines](/agntcy/repo-template/blob/main/CONTRIBUTING.md)
- [x] Existing issues have been referenced (where applicable)
- [x] I have verified this change is not present in other open pull
requests
- [x] Functionality is documented
- [x] All code style checks pass
- [x] New code contribution is covered by automated tests
- [x] All new and existing tests pass

Signed-off-by: Michele Papalini <micpapal@cisco.com>