fix(session): remove participants from the group list (#1059)
# Description

Fixes a bug where participants are not properly removed from the group
list in a multicast session with only 2 participants (moderator + 1
client).

## Problem

In a multicast session with only 2 participants (moderator and one
client), when the moderator removes the participant sending a
LeaveRequest, the participant was not removed from the group_list in the
control message sender. This caused:

- The moderator to maintain stale participant state
- The moderator was potentially waiting for acks for messages from
removed participants
- The missing_pings map to accumulate entries for participants who
already left

Fix: #1058 

## Type of Change

- [x] Bugfix
- [ ] New Feature
- [ ] Breaking Change
- [ ] Refactor
- [ ] Documentation
- [ ] Other (please describe)

## Checklist

- [x] I have read the [contributing
guidelines](/agntcy/repo-template/blob/main/CONTRIBUTING.md)
- [x] Existing issues have been referenced (where applicable)
- [x] I have verified this change is not present in other open pull
requests
- [x] Functionality is documented
- [x] All code style checks pass
- [x] New code contribution is covered by automated tests
- [x] All new and existing tests pass

---------

Signed-off-by: Michele Papalini <micpapal@cisco.com>